fix: 구간 관리 페이지의 모달이 store 상태 변경에 따라 매번 리렌더링 되던 버그를 수정
