+++ src/components/SectionsDetail/index.jsx
@@ -1,36 +1,30 @@
 import React from "react";
 import PropTypes from "prop-types";
 import cx from "classnames";
+import { useSelector } from "react-redux";
+import { selectLineByLineId } from "../../pages/Lines/slice";
 import SectionsDetailList from "../SectionsDetailList";
 
-const SectionsDetail = ({ detail }) => {
-  const { name, color, stations, id } = detail ?? {};
+const SectionsDetail = ({ lineId }) => {
+  const line = useSelector((state) => selectLineByLineId(state, lineId));
 
-  return (
-    detail && (
-      <div className="border rounded-md">
-        <h3
-          className={cx("pl-4 py-2 text-gray-800 text-xl rounded-t-md", color)}
-        >
-          {name}
-        </h3>
-        <SectionsDetailList stations={stations} lineId={id} />
-      </div>
-    )
+  return !line ? null : (
+    <div className="border rounded-md">
+      <h3
+        className={cx(
+          "pl-4 py-2 text-gray-800 text-xl rounded-t-md",
+          line.color
+        )}
+      >
+        {line.name}
+      </h3>
+      <SectionsDetailList stations={line.stations} lineId={lineId} />
+    </div>
   );
 };
 
 SectionsDetail.propTypes = {
-  detail: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    color: PropTypes.string.isRequired,
-    stations: PropTypes.string.isRequired,
-  }),
-};
-
-SectionsDetail.defaultProps = {
-  detail: null,
+  lineId: PropTypes.string.isRequired,
 };
 
 export default SectionsDetail;
